refactor: MainCarouselView를 담는 CarouselCollectionReusableView 구현
